+++ src/utils/axiosSingleReview.js
@@ -1,0 +1,9 @@
+import { ApiRequest } from "./axios";
+
+export const getReviewById = (review_id) => {
+  return ApiRequest.get(`/reviews/${review_id}`).then(
+    ({ data: { review } }) => {
+      return review;
+    }
+  );
+};
